Task FP.1
- Define a class TrafficLight which is a child class of TrafficObject.
The class shall have the public methods void waitForGreen() and
void simulate() as well as TrafficLightPhase getCurrentPhase(), where
TrafficLightPhase is an enum that can be either red or green. Also, add
the private method void cycleThroughPhases(). Furthermore, there shall
be the private member _currentPhase which can take red or green as its
value.